requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
redis==4.5.5
types-pyOpenSSL==23.1.0.3
types-redis==4.5.5.2
typing_extensions==4.6.2
yarl==1.9.2
